Last Version
A program that measures how much time individuals spend around a desk in a workspace. Regions of interest (ROIs) are used to detect the areas surrounding the desk. The time library is used to measure the duration spent in the designated regions. People are detected using OpenCV. As a result, the productivity within the workspace can be monitored. Due to privacy concerns, videos and photos cannot be shared.